requirements.txt
opencv-python==4.10.0.84 
# opencv-contrib-python==4.5.5.64
opencv-contrib-python==4.10.0.84 
tensorflow==2.12.0
numpy==1.23.5 
pandas==2.2.3
Pillow==9.4.0
yagmail==0.15.285